consolidate ab test helper functions
