+++ assets/helpers/abTests/abtest.js
@@ -202,6 +202,17 @@ const getVariantsAsString = (participation: Participations): string => {
 
 const getCurrentParticipations = (): Participations => getLocalStorageParticipation();
 
+function getVariantForTest(participations: Participations, test: string) {
+  return participations ? participations[test] : null;
+}
+
+function getPaymentLogosTestVariant(participations: Participations) {
+  return getVariantForTest(participations, 'paymentLogosTest');
+}
+
+function inPaymentLogosTest(participations: Participations) {
+  return getPaymentLogosTestVariant(participations) === 'variant';
+}
 
 // ----- Exports ----- //
 
@@ -209,4 +220,6 @@ export {
   init,
   getVariantsAsString,
   getCurrentParticipations,
+  getPaymentLogosTestVariant,
+  inPaymentLogosTest,
 };
