+++ assets/pages/contributions-landing/containers/payPalContributionButtonContainer.js
@@ -7,7 +7,7 @@ import { connect } from 'react-redux';
 import PayPalContributionButton from 'containerisableComponents/payPalContributionButton/payPalContributionButton';
 import { payPalContributionButtonActionsFor } from 'containerisableComponents/payPalContributionButton/payPalContributionButtonActions';
 import { getAmount } from 'containerisableComponents/contributionSelection/contributionSelectionReducer';
-import { inPaymentLogosTest } from 'helpers/abTests/helpers';
+import { inPaymentLogosTest } from 'helpers/abTests/abtest';
 import type { State } from '../contributionsLandingReducer';
